Whole program is restructurized - previous topology_prep was partitioned into two files: initial_struc.py and topology_prep.py
